--- src/@noctua-doctor/data/obsolete.js
@@ -1,3 +1,0 @@
-export const obsolete = {
-
-}
