--- src/@noctua-doctor/models/menu-panels.ts
@@ -1,24 +1,0 @@
-export enum LeftPanel {
-    home = 'home',
-    checklist = 'checklist',
-    obsolete = 'obsolete',
-    terms = 'terms',
-    group = 'group',
-    contributor = 'contributor',
-    calendar = 'calendar',
-    chat = 'chat',
-    notes = 'notes',
-    activities = 'activities'
-};
-
-export enum MiddlePanel {
-    home = 'home',
-    cams = 'cams',
-    camsReview = 'camReview',
-    reviewChanges = 'reviewChanges'
-};
-
-export enum RightPanel {
-    camForm = 'camForm',
-    copyModel = 'copyModel',
-}
