+++ src/@noctua.doctor/data/obsolete.js
@@ -1,0 +1,3 @@
+export const obsolete = {
+
+}
