+++ src/app/layout/components/footer/footer.component.html
@@ -13,7 +13,7 @@
   <span fxFlex=""></span>
   <div class="" style="text-align: right;" fxFlex="40%">
     <small class="">
-      Copyright © 1999-2022
+      Copyright © 1999-2020
       Member of the
       <a href=" http://www.obofoundry.org/" class="footer__link" rel="external" title="Open Biological Ontologies"
         target="blank">Open Biological Ontologies Foundry</a>
