+++ src/app/main/apps/noctua-search/cams/cams-table/cams-table.component.ts
@@ -114,7 +114,6 @@ export class CamsTableComponent implements OnInit, OnDestroy {
       'date',
       'contributor',
       'edit',
-      'export',
     ];
 
     if (isReviewMode) {
