+++ workbenches/noctua-landing-page/public/inject.tmpl
@@ -214,7 +214,6 @@
     </div>
   </noctua-splash-screen>
   <noctua-root></noctua-root>
-<script src="runtime.38b1f495cb4cd82a.js" type="module"></script><script src="polyfills.a39db01cea6f891e.js" type="module"></script><script src="scripts.79b3808e7d5e20c8.js" defer></script><script src="main.4be18bb03e6ecdb9.js" type="module"></script>
-
+<script src="runtime.38b1f495cb4cd82a.js" type="module"></script><script src="polyfills.a39db01cea6f891e.js" type="module"></script><script src="scripts.79b3808e7d5e20c8.js" defer></script><script src="main.792c2c4a3cb3b99c.js" type="module"></script>
 
 </body></html>
